fix: iframe for numeri page
